+++ src/server/dcp_server/utils/fsimagestorage.py
